pari is now linked in:
    i11 : 3483856828159805266414750134638693 * 40201127129528507444587

    o11 = 140054971249928282912580794066400368104344406829963604791

    i12 : time Pari$factor oo
	 -- used 17.6811 seconds

    o12 = ((40201127129528507444587, 3483856828159805266414750134638693), (1, 1))

    o12 : Sequence

    i13 : times apply(first oo, last oo, power)

    o13 = 140054971249928282912580794066400368104344406829963604791



git-svn-id: svn://svn.macaulay2.com/Macaulay2/trunk/M2@8009 a8012f80-5f13-0410-8dc0-eb77bd940c0d

+++ Macaulay2/d/actors5.d
@@ -622,7 +622,7 @@ setupfun("unstack",unstack);
 alarm(e:Expr):Expr := (
      when e is i:ZZ do 
      if isInt(i)
-     then Expr(toInteger(alarm(toInt(i))))
+     then Expr(toInteger(int(alarm(uint(toInt(i))))))
      else WrongArgSmallInteger()
      else WrongArgZZ()
      );
